Update index.html
In this updated code, I added a CSS class called .experience-container to the section element containing the job experiences. The .experience-container class is set to display: flex; and justify-content: space-between;, which aligns the job sections to the left and right sides of the container.

I also added a .job class to each job's div element inside the experience-container. The .job class is set to width: 45%; to ensure that each job takes up 45% of the available space in the container, leaving some space between the job sections.